Major change in the documentation's format
Standarized the documentation, so it now looks more similar across
documents, changed the formating for titles and and document structures
so that now a single .html file can be generated and/or a single pdf
file can be generated. The intention for this, is to have the
information available at http://open.qa/documentation (some work also
needs to be done there to match the changes here.

There is still some work to be done regading internal links on the
output files (e.g converting with asciidoctor or asciidoctor-pdf).

The index.asciidoc file has the sole purpose of having everything
combined to generate a single file (HTML or PDF).

Fix rendering problems and change Authors directive

Had to change the formatting a bit of few documents so that the actual
rendering of the "final" documents is properly shown either on PDF or
HTML, on github it looks still fine.

Change Authors to OpenQA Team :)