Refactor Cache to use SQLite instead of plain json
With the problems we've with the current implementation of the cache
the asset locking is not working as expected, leading to workers not
being able to keep downloading files, once they hit a race condition
with the lock files.

This refactor attempts to solve the situation, by using sqlite, since
it is able to handle on it's own database locking and proper
transactions.

In case the cache.db file is not present, it will be deployed from
scratch.

Since we are saving now size and etag, we can now easily know if an
asset has been changed, by sending the etag during the first request
and bailing out if there are no changes