Use Mojo::IOLoop for monitoring
Trigger the monitoring of the master WebUI and slave WebUI, and
use events to handle new data.

Initially being set up as stubs but the bridge should be able to find
new items on the WebUI that are scheduled for it, and transfer
them to the newer slave, after changing the job settings.